Remove artificial test_havoc_deadlock and add loom model
